+++ src/main.tsx
@@ -1,14 +1,24 @@
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-//import App from "app/app";
 import Home from "./pages/home";
+import System from "./pages/system";
 import Layout from "components/layout";
+
+export function App() {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<Layout><Home /></Layout>} />
+        <Route path="/system" element={<Layout><System /></Layout>} />
+      </Routes>
+    </Router>
+  );
+}
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <Layout>
-      <Home />
-    </Layout>
+    <App />
   </React.StrictMode>,
 );