+++ src/pages/home/index.tsx
@@ -1,8 +1,7 @@
-
 export default function Home() {
     return (
         <div>
-            <p>HELLO THIS IS TESTING HAHAHA</p>
+            <p>WELCOME</p>
         </div>
     );
 }