+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-# 1.0.1
+# Dev Update 1.0.2
+- Testing pages
+- Testing components
+- Added routes
+- Added logos
+- Added typeORM for handling database communication
+- Added .env stuff
+- Added navbar component
+
+# Dev Update 1.0.1
 - Setting up the basic layout and starting pages.
 - Added a changelog for personal reference.